# By Alexey Strokachuk (6) and LeRoN (1)
# Via Alexey Strokachuk
* commit '177ca11fa1752110eb07cfbfd7ded8bf1cda37a5 [formerly 57bf0ad48cd705b723f5d775a78ca295bd7bb0d5]':
  [DF-6922] fix compilation for win10
  [DF-6922] [DF-8229] [DF-8304] [DF-6516] fix macOs project
  [DF-6922] [DF-8229] [DF-8304] [DF-6516] fix enum name. fix redundant windows on Win32, fix assert under MacOS
  [DF-6922] remove legacy code from resource editor.
  [DF-6922] fix review issues.
  [DF-6922] fix iOS&MacOS project
  [DF-6922] Add legacy controls in QuickEd. Remove UIHierarchy/UIHierarchyCell/UIHierarchyNode. Removed some legacy methods from UIControl.


Former-commit-id: e8db32d6e496b2ad72677300c6aa7eec3a8032ed

+++ Tools/QuickEd/Classes/Model/ControlProperties/AbstractProperty.cpp
@@ -50,12 +50,12 @@ void AbstractProperty::SetParent(AbstractProperty *parent)
     this->parent = parent;
 }
 
-int AbstractProperty::GetIndex(AbstractProperty *property) const
+int32 AbstractProperty::GetIndex(AbstractProperty *property) const
 {
-    for (int32 i = 0; i < GetCount(); i++)
+    for (uint32 i = 0; i < GetCount(); i++)
     {
         if (GetProperty(i) == property)
-            return i;
+            return (int32)i;
     }
     return -1;
 }
@@ -66,7 +66,7 @@ void AbstractProperty::Refresh(DAVA::int32 refreshFlags)
 
 AbstractProperty *AbstractProperty::FindPropertyByPrototype(AbstractProperty *prototype)
 {
-    for (int32 i = 0; i < GetCount(); i++)
+    for (uint32 i = 0; i < GetCount(); i++)
     {
         AbstractProperty *result = GetProperty(i)->FindPropertyByPrototype(prototype);
         if (result)
@@ -77,7 +77,7 @@ AbstractProperty *AbstractProperty::FindPropertyByPrototype(AbstractProperty *pr
 
 bool AbstractProperty::HasChanges() const
 {
-    for (int32 i = 0; i < GetCount(); i++)
+    for (uint32 i = 0; i < GetCount(); i++)
     {
         if (GetProperty(i)->HasChanges())
             return true;
